+++ src/ThreeDAnalyzer.Web/obj/Debug/net10.0/ThreeDAnalyzer.Web.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+8e55fc34c83ca689f96ed27ebc29b122d224fb67")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+71d3e3b3ac8de44e7a18cb94a92eaaf5ffa87e95")]
 [assembly: System.Reflection.AssemblyProductAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]